+++ CHANGELOG.md
@@ -1,9 +1,18 @@
 # Changelog
 All notable changes to this project will be documented in this file.
 
-The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
+The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
 
-[Unreleased]: https://github.com/fairdevkit/rdf-resource-resolver/compare/master...HEAD
+
+## [0.2.0] - 2024 Oct 28
+
+### Changed (backward incompatible)
+
+- changed package namespace from `nl.dtls` to `org.fairdatateam`
+- changed `groupId` to `org.fairdatateam.rdf`
+
+[Unreleased]: https://github.com/fairdatateam/rdf-resource-resolver/compare/master...HEAD
+[0.2.0]: https://github.com/fairdatateam/rdf-resource-resolver-fork/releases/tag/0.2.0